Integration tests for java_gradle
