+++ tests/integration/testdata/buildcmd/Java/gradle/build.gradle
@@ -8,4 +8,7 @@ repositories {
 
 dependencies {
     implementation 'software.amazon.awssdk:annotations:2.1.0'
+	compile (
+        'com.amazonaws:aws-lambda-java-core:1.1.0'
+    )
 }
