--- tests/integration/testdata/buildcmd/Java/gradle/src/main/java/aws/lambdabuilders/Main.java
@@ -1,7 +1,0 @@
-package aws.lambdabuilders;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello AWS Lambda Builders!");
-    }
-}
